+++ livre/lib/configuration/VolumeRendererParameters.cpp
@@ -18,7 +18,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include <livre/lib/configuration/VolumeRendererParameters.h>
+#include "VolumeRendererParameters.h"
 
 namespace livre
 {
@@ -31,111 +31,55 @@ const std::string MINLOD_PARAM = "min-lod";
 const std::string MAXLOD_PARAM = "max-lod";
 const std::string SAMPLESPERRAY_PARAM = "samples-per-ray";
 const std::string SAMPLESPERPIXEL_PARAM = "samples-per-pixel";
-const std::string TRANSFERFUNCTION_PARAM = "transfer-function";
 
 VolumeRendererParameters::VolumeRendererParameters()
     : Parameters( "Volume Renderer Parameters" )
-    , synchronousMode( false )
-#ifdef __i386__
-    , screenSpaceError( 8.0f )
-    , maxGPUCacheMemoryMB( 384u )
-    , maxCPUCacheMemoryMB( 768u )
-#else
-    , screenSpaceError( 4.0f )
-    , maxGPUCacheMemoryMB( 3072u )
-    , maxCPUCacheMemoryMB( 8192u )
-#endif
-    , minLOD( 0 )
-    , maxLOD( ( NODEID_LEVEL_BITS << 1 ) + 1 )
-    , samplesPerRay( 0 )
-    , samplesPerPixel( 1u )
-    , transferFunction()
 {
+#ifdef __i386__
+    setSSE( 8.0f );
+    setMaxGPUCacheMemoryMB( 384u );
+    setMaxCPUCacheMemoryMB( 768u );
+#endif
+
     configuration_.addDescription( configGroupName_, GPUCACHEMEM_PARAM,
                                    "Maximum GPU cache memory (MB) - "
                                    "caches the texture data in GPU memory",
-                                   maxGPUCacheMemoryMB );
+                                   getMaxGPUCacheMemoryMB( ));
     configuration_.addDescription( configGroupName_, CPUCACHEMEM_PARAM,
                                    "Maximum CPU cache memory (MB) - "
                                    "caches the volume data in CPU memory",
-                                   maxCPUCacheMemoryMB );
+                                   getMaxCPUCacheMemoryMB( ));
     configuration_.addDescription( configGroupName_, SCREENSPACEERROR_PARAM,
-                                   "Screen space error", screenSpaceError );
+                                   "Screen space error", getSSE( ));
     configuration_.addDescription( configGroupName_, SYNCHRONOUSMODE_PARAM,
-                                   "Enable synchronous mode", synchronousMode );
+                                   "Enable synchronous mode", getSynchronousMode( ));
     configuration_.addDescription( configGroupName_, MINLOD_PARAM,
-                                   "Minimum level of detail", minLOD );
+                                   "Minimum level of detail", getMinLOD( ));
     configuration_.addDescription( configGroupName_, MAXLOD_PARAM,
-                                   "Maximum level of detail", maxLOD );
+                                   "Maximum level of detail", getMaxLOD( ));
     configuration_.addDescription( configGroupName_, SAMPLESPERRAY_PARAM,
                                    "Number of samples per ray. The value of 0 (default) sets"
                                    " the number of samples automatically, based on the volume size",
-                                   samplesPerRay );
+                                   getSamplesPerRay( ));
     configuration_.addDescription( configGroupName_, SAMPLESPERPIXEL_PARAM,
-                                   "Number of samples per pixel", samplesPerPixel );
-    configuration_.addDescription( configGroupName_, TRANSFERFUNCTION_PARAM,
-                                ".1dt transfer function file (from ImageVis3D)",
-                                   transferFunction );
-}
-
-void VolumeRendererParameters::deserialize( co::DataIStream& is, const uint64_t )
-{
-    is >> screenSpaceError
-       >> synchronousMode
-       >> maxGPUCacheMemoryMB
-       >> maxCPUCacheMemoryMB
-       >> minLOD
-       >> maxLOD
-       >> samplesPerRay
-       >> samplesPerPixel
-       >> transferFunction;
-}
-
-void VolumeRendererParameters::serialize( co::DataOStream& os, const uint64_t )
-{
-    os << screenSpaceError
-       << synchronousMode
-       << maxGPUCacheMemoryMB
-       << maxCPUCacheMemoryMB
-       << minLOD
-       << maxLOD
-       << samplesPerRay
-       << samplesPerPixel
-       << transferFunction;
-}
-
-VolumeRendererParameters& VolumeRendererParameters::operator=(
-        const VolumeRendererParameters& rhs )
-{
-    if( this == &rhs )
-        return *this;
-
-    screenSpaceError = rhs.screenSpaceError;
-    synchronousMode = rhs.synchronousMode;
-    maxGPUCacheMemoryMB = rhs.maxGPUCacheMemoryMB;
-    maxCPUCacheMemoryMB = rhs.maxCPUCacheMemoryMB;
-    minLOD = rhs.minLOD;
-    maxLOD = rhs.maxLOD;
-    samplesPerRay = rhs.samplesPerRay;
-    samplesPerPixel = rhs.samplesPerPixel;
-    transferFunction = rhs.transferFunction;
-    setDirty( DIRTY_ALL );
-
-    return *this;
+                                   "Number of samples per pixel", getSamplesPerPixel( ));
 }
 
 void VolumeRendererParameters::initialize_()
 {
-    configuration_.getValue( SYNCHRONOUSMODE_PARAM, synchronousMode );
-    configuration_.getValue( SCREENSPACEERROR_PARAM, screenSpaceError );
-    configuration_.getValue( GPUCACHEMEM_PARAM, maxGPUCacheMemoryMB );
-    configuration_.getValue( CPUCACHEMEM_PARAM, maxCPUCacheMemoryMB);
-    configuration_.getValue( MINLOD_PARAM, minLOD );
-    configuration_.getValue( MAXLOD_PARAM, maxLOD );
-    configuration_.getValue( SAMPLESPERRAY_PARAM, samplesPerRay );
-    configuration_.getValue( SAMPLESPERPIXEL_PARAM, samplesPerPixel );
-    configuration_.getValue( TRANSFERFUNCTION_PARAM, transferFunction );
-    setDirty( DIRTY_ALL );
+    setSynchronousMode( configuration_.getValue( SYNCHRONOUSMODE_PARAM,
+                                                 getSynchronousMode( )));
+    setSSE( configuration_.getValue( SCREENSPACEERROR_PARAM, getSSE( )));
+    setMaxGPUCacheMemoryMB( configuration_.getValue( GPUCACHEMEM_PARAM,
+                                                     getMaxGPUCacheMemoryMB( )));
+    setMaxCPUCacheMemoryMB( configuration_.getValue( CPUCACHEMEM_PARAM,
+                                                     getMaxCPUCacheMemoryMB( )));
+    setMinLOD( configuration_.getValue( MINLOD_PARAM, getMinLOD( )));
+    setMaxLOD( configuration_.getValue( MAXLOD_PARAM, getMaxLOD( )));
+    setSamplesPerRay( configuration_.getValue( SAMPLESPERRAY_PARAM,
+                                               getSamplesPerRay( )));
+    setSamplesPerPixel( configuration_.getValue( SAMPLESPERPIXEL_PARAM,
+                                                 getSamplesPerPixel( )));
 }
 
 } //Livre